float compatibility+ example

+++ shapes/__init__.py
@@ -6,11 +6,11 @@ class rectangle():
         self.width = width
 
     def area(self):
-        area_op = int(self.height)*int(self.width)
+        area_op = float(self.height)*float(self.width)
         return area_op
 
     def perimeter(self):
-        perimeter = 2*(int(self.height) + int(self.width))
+        perimeter = 2*(float(self.height) + float(self.width))
         return perimeter
 
 class square():
@@ -18,11 +18,11 @@ class square():
         self.length = length
     
     def area(self):
-        area = int(self.length)**2
+        area = float(self.length)**2
         return area
 
     def perimeter(self):
-        perimeter = 4*int(self.length)
+        perimeter = 4*float(self.length)
         return perimeter
 
 class circle():
@@ -30,9 +30,9 @@ class circle():
         self.radius = radius
 
     def perimeter(self):
-        perimeter = 2*math.pi*int(self.radius)
+        perimeter = 2*math.pi*float(self.radius)
         return perimeter
 
     def area(self):
-        area = math.pi*int(self.radius)**2
+        area = math.pi*float(self.radius)**2
         return area
